requirements.txt
alembic==1.13.1
asgiref==3.7.2
astroid==3.1.0
bcrypt==4.1.2
Bcrypt-Flask==1.0.2
black==24.4.0
blinker==1.7.0
boto3==1.34.72
botocore==1.34.72
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dill==0.3.8
distlib==0.3.7
dnspython==2.6.1
email_validator==2.1.1
filelock==3.12.4
flake8==7.0.0
Flask==3.0.3
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-CKEditor==0.4.5
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.0
Flask-Modals==0.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.4
jmespath==1.0.1
lazy-object-proxy==1.10.0
Mako==1.3.5
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pillow==10.2.0
platformdirs==3.10.0
pluggy==1.4.0
psycopg2==2.9.9
pycodestyle==2.11.1
pycparser==2.22
pycryptodome==3.20.0
pyflakes==3.2.0
pylint==3.1.0
pytest==8.1.1
python-dateutil==2.9.0.post0
requests==2.31.0
s3transfer==0.10.1
secret==0.8
six==1.16.0
SQLAlchemy==2.0.29
sqlparse==0.4.4
tabulate==0.9.0
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.10.0
tzdata==2023.3
urllib3==2.2.1
virtualenv==20.24.5
Werkzeug==3.0.1
wrapt==1.16.0
WTForms==3.1.2
WTForms-SQLAlchemy==0.4.1
